// src/app/core/models/register.model.ts
export class Register {
  firstName: string = '';
  lastName: string = '';
  email: string = '';
  password: string = '';
  confirmPwd: string = '';
  gender: string = '';
  month: string = '';
  day: string = '';
  year: string = '';
  code: string = '';
  profileType: string = '';
}